LocalHandle incorporated into LocalInstance
