+++ src/main/kotlin/com/cognifide/gradle/aem/common/Formats.kt
@@ -59,18 +59,18 @@ object Formats {
     }
 
     fun size(file: File): String {
-        return bytesToHuman(FileUtils.sizeOf(file))
+        return bytesToHuman(when {
+            file.exists() -> FileUtils.sizeOf(file)
+            else -> 0L
+        })
     }
 
     fun bytesToHuman(bytes: Long): String {
-        if (bytes < 1024) {
-            return bytes.toString() + " B"
-        } else if (bytes < 1024 * 1024) {
-            return (bytes / 1024).toString() + " KB"
-        } else if (bytes < 1024 * 1024 * 1024) {
-            return String.format("%.2f MB", bytes / (1024.0 * 1024.0))
-        } else {
-            return String.format("%.2f GB", bytes / (1024.0 * 1024.0 * 1024.0))
+        return when {
+            bytes < 1024 -> bytes.toString() + " B"
+            bytes < 1024 * 1024 -> (bytes / 1024).toString() + " KB"
+            bytes < 1024 * 1024 * 1024 -> String.format("%.2f MB", bytes / (1024.0 * 1024.0))
+            else -> String.format("%.2f GB", bytes / (1024.0 * 1024.0 * 1024.0))
         }
     }
 
@@ -79,7 +79,11 @@ object Formats {
     }
 
     fun percent(current: Long, total: Long): String {
-        val value: Double = if (total == 0L) 0.0 else current.toDouble() / total.toDouble()
+        val value: Double = when (total) {
+            0L -> 0.0
+            else -> current.toDouble() / total.toDouble()
+        }
+
         return "${"%.2f".format(value * 100.0)}%"
     }
