+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tasks/Instance.kt
@@ -1,23 +1,17 @@
 package com.cognifide.gradle.aem.instance.tasks
 
 import com.cognifide.gradle.aem.common.AemDefaultTask
-import com.cognifide.gradle.aem.instance.LocalHandle
 import com.cognifide.gradle.aem.instance.LocalInstance
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.Internal
 
 open class Instance : AemDefaultTask() {
 
     @Input
-    var localHandles: List<LocalHandle> = listOf()
-
-    @get:Internal
-    val instances: List<LocalInstance>
-        get() = localHandles.map { it.instance }
+    var instances: List<LocalInstance> = listOf()
 
     override fun projectsEvaluated() {
-        if (localHandles.isEmpty()) {
-            localHandles = aem.localHandles
+        if (instances.isEmpty()) {
+            instances = aem.localInstances
         }
     }
 }
