+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tasks/Up.kt
@@ -1,6 +1,6 @@
 package com.cognifide.gradle.aem.instance.tasks
 
-import com.cognifide.gradle.aem.instance.LocalHandle
+import com.cognifide.gradle.aem.instance.LocalInstance
 import com.cognifide.gradle.aem.instance.action.AwaitAction
 import com.cognifide.gradle.aem.instance.names
 import org.gradle.api.tasks.TaskAction
@@ -11,14 +11,14 @@ open class Up : Instance() {
         description = "Turns on local AEM instance(s)."
     }
 
-    private var initOptions: LocalHandle.() -> Unit = {}
+    private var initOptions: LocalInstance.() -> Unit = {}
 
     private var awaitOptions: AwaitAction.() -> Unit = {}
 
     /**
      * Hook called only when instance is up first time.
      */
-    fun init(options: LocalHandle.() -> Unit) {
+    fun init(options: LocalInstance.() -> Unit) {
         this.initOptions = options
     }
 
@@ -31,16 +31,16 @@ open class Up : Instance() {
 
     @TaskAction
     fun up() {
-        aem.parallelWith(localHandles) { up() }
+        aem.parallelWith(instances) { up() }
 
         aem.actions.await {
             instances = this@Up.instances
             awaitOptions()
         }
 
-        aem.parallelWith(localHandles) { init(initOptions) }
+        aem.parallelWith(instances) { init(initOptions) }
 
-        aem.notifier.notify("Instance(s) up", "Which: ${localHandles.names}")
+        aem.notifier.notify("Instance(s) up", "Which: ${instances.names}")
     }
 
     companion object {
